Pull common vec code out into module
