+++ src/compiler/typescript_to_rust/azle_vm_value_derive/src/lib.rs
@@ -8,6 +8,7 @@ mod derive_try_into_vm_value {
 mod derive_try_from_vm_value {
     pub mod derive_try_from_vm_value_enum;
     pub mod derive_try_from_vm_value_struct;
+    pub mod derive_try_from_vm_value_vec;
 }
 
 use derive_try_from_vm_value::{
